requirements.txt
python-http-client==3.3.7
sendgrid==6.11.0
starkbank-ecdsa==2.2.0
azure-functions==1.13.3